+++ CHANGELOG.md
@@ -11,9 +11,11 @@ and this project adheres to
 ### Added
 
 - Add fallback track support
-- Fetch track details in separated threads
+- Fetch track details in separated threads (when requested)
 - Add the `CONNECTION_POOL_MAXSIZE` configuration setting
-- CLI: add track release date to outputs
+- Add the `ALWAYS_FETCH_RELEASE_DATE` configuration setting
+- CLI: add track/album release date to outputs for the `search` 
+  and `artist` commands using the `-r/--release` option
 
 ## [0.5.2] - 2025-11-04